made some additions the the LearningAgent logic
