+++ src/main/java/com/maastricht/university/logic/ai/hybrid/environment/Environment.java
@@ -1,12 +1,10 @@
 package com.maastricht.university.logic.ai.hybrid.environment;
 
 import com.maastricht.university.logic.ai.agent.Agent;
-import com.maastricht.university.logic.ai.agent.LearningAgent;
+import com.maastricht.university.logic.ai.hybrid.LearningAgent;
 import com.maastricht.university.logic.ai.agent.RandomAgent;
-import com.maastricht.university.logic.ai.minimax.tree.ITreeNode;
 import com.maastricht.university.logic.ai.reinforcement.network.RewardCalculator;
 import com.maastricht.university.logic.game.game.GameState;
-import com.maastricht.university.logic.game.util.interfaces.IGameState;
 import org.deeplearning4j.gym.StepReply;
 import org.deeplearning4j.rl4j.mdp.MDP;
 import org.deeplearning4j.rl4j.space.DiscreteSpace;
@@ -24,11 +22,20 @@ public class Environment implements MDP<NeuralGameState, Integer, DiscreteSpace>
     @Override
     public StepReply<NeuralGameState> step(Integer integer) {
         // TODO: make this step function iteratively set the score of leaves of a minimax tree.
-        ITreeNode<IGameState> node = learningAgent.getNextLeaf();
-        node.setScore(integer);
+
+        GameState nextState;
+        boolean isDoneComputingTree = !learningAgent.isNull();
+
+        if (isDoneComputingTree) {
+            learningAgent.move();
+            nextState = game;
+        } else {
+            learningAgent.getCurrentNode().setScore(integer);
+            nextState = (GameState) learningAgent.getNextNode().getElement();
+        }
 
         return new StepReply<>(new NeuralGameState(
-                ((GameState)node.getElement()).getStateVector()),
+                nextState.getStateVector()),
                 RewardCalculator.getWinReward(game, 1),
                 isDone(),
                 null
